fx_feed/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fx-feed"
version = "0.1.0"
description = "FX bookticker 수집 (Binance / dxFeed / Databento) → QuestDB. Gate HFT 전략 reference 후보 탐색용."
requires-python = ">=3.10"
dependencies = [
    "websockets>=13.0",
    "httpx>=0.27",
    "uvloop>=0.21; sys_platform != 'win32'",
    "structlog>=24.0",
    "tomli>=2.0; python_version < '3.11'",
    "questdb>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dxfeed = []
databento = [
    "databento>=0.45",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.10",
]

[project.scripts]
fx-feed = "fx_feed.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fx_feed"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
